Убрал компонент canvas добавил таймер и подсчет очков
