+++ src/classes/index.ts
@@ -1,0 +1,8 @@
+import Asteroid from './Asteroid/Asteroid';
+import Ship from './Ship/Ship';
+import Bullet from './Bullet/Bullet';
+import Particle from './Particle/Particle';
+
+export {
+  Asteroid, Ship, Bullet, Particle,
+};
